[CHORE]Move all styles under styles folder
